+++ core/src/fr/hammons/slinc/modules/FSetModule.scala
@@ -1,12 +1,12 @@
 package fr.hammons.slinc.modules
 
 import fr.hammons.slinc.FSetBacking
-import fr.hammons.slinc.CFunctionBindingGenerator
+import fr.hammons.slinc.FunctionBindingGenerator
 import fr.hammons.slinc.CFunctionDescriptor
 
 trait FSetModule:
   val runtimeVersion: Int
   def getBacking(
       desc: List[CFunctionDescriptor],
-      generators: List[CFunctionBindingGenerator]
+      generators: List[FunctionBindingGenerator]
   ): FSetBacking[?]
